test: add test-dns-promises-lookupService
This adds covereage for the onlookupservice() callback in
lib/internal/dns/promises.js. Because of stubbing in other tests, it is
not currently covered.

This test works on my local development machine with the network turned
off, so I'm putting it in parallel. If CI proves more challenging, it
can be moved to the internet directory instead.

PR-URL: https://github.com/nodejs/node/pull/31640
Reviewed-By: Colin Ihrig <cjihrig@gmail.com>
Reviewed-By: James M Snell <jasnell@gmail.com>
Reviewed-By: Luigi Pinca <luigipinca@gmail.com>
Reviewed-By: Anna Henningsen <anna@addaleax.net>
Reviewed-By: Richard Lau <riclau@uk.ibm.com>